Move implementation to shared BaseViewModel
