+++ shared/src/androidMain/kotlin/com/agilefreaks/sharedappsample/PlatformBaseViewModel.kt
@@ -1,58 +1,13 @@
 package com.agilefreaks.sharedappsample
 
-import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.State
-import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.flow.*
-import kotlinx.coroutines.launch
 
-actual abstract class PlatformBaseViewModel<Event : ViewEvent, UiState : ViewState, Effect : ViewSideEffect> :
+actual abstract class PlatformBaseViewModel :
     ViewModel() {
 
     actual val scope = viewModelScope
 
-    private val initialState: UiState by lazy { setInitialState() }
-    abstract fun setInitialState(): UiState
-
-    private val _viewState: MutableState<UiState> = mutableStateOf(initialState)
-    val viewState: State<UiState> = _viewState
-
-    private val _event: MutableSharedFlow<Event> = MutableSharedFlow()
-
-    private val _effect: Channel<Effect> = Channel()
-    val effect = _effect.receiveAsFlow()
-
-    init {
-        subscribeToEvents()
-    }
-
     actual override fun onCleared() {
-    }
-
-    fun setEvent(event: Event) {
-        scope.launch { _event.emit(event) }
-    }
-
-    protected fun setState(reducer: UiState.() -> UiState) {
-        val newState = viewState.value.reducer()
-        _viewState.value = newState
-    }
-
-    private fun subscribeToEvents() {
-        scope.launch {
-            _event.collect {
-                handleEvents(it)
-            }
-        }
-    }
-
-    abstract fun handleEvents(event: Event)
-
-    protected fun setEffect(builder: () -> Effect) {
-        val effectValue = builder()
-        scope.launch { _effect.send(effectValue) }
     }
 }
