+++ shared/src/commonMain/kotlin/com/agilefreaks/sharedappsample/BaseViewModel.kt
@@ -1,6 +1,12 @@
 package com.agilefreaks.sharedappsample
 
 import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.flow.MutableSharedFlow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.receiveAsFlow
+import kotlinx.coroutines.launch
 
 interface ViewState
 
@@ -8,14 +14,54 @@ interface ViewEvent
 
 interface ViewSideEffect
 
-expect abstract class PlatformBaseViewModel<Event : ViewEvent, UiState : ViewState, Effect : ViewSideEffect>() {
+expect abstract class PlatformBaseViewModel() {
     val scope: CoroutineScope
 
     protected open fun onCleared()
 }
 
 abstract class BaseViewModel<Event : ViewEvent, UiState : ViewState, Effect : ViewSideEffect>:
-    PlatformBaseViewModel<Event, UiState, Effect>() {
+    PlatformBaseViewModel() {
+
+    private val initialState: UiState by lazy { setInitialState() }
+    abstract fun setInitialState(): UiState
+
+    private val _viewState: MutableStateFlow<UiState> = MutableStateFlow(initialState)
+    val viewState: StateFlow<UiState> = _viewState
+
+    private val _event: MutableSharedFlow<Event> = MutableSharedFlow()
+
+    private val _effect: Channel<Effect> = Channel()
+    val effect = _effect.receiveAsFlow()
+
+    init {
+        subscribeToEvents()
+    }
+
+    fun setEvent(event: Event) {
+        scope.launch { _event.emit(event) }
+    }
+
+    protected fun setState(reducer: UiState.() -> UiState) {
+        scope.launch {
+            _viewState.emit(viewState.value.reducer())
+        }
+    }
+
+    private fun subscribeToEvents() {
+        scope.launch {
+            _event.collect {
+                handleEvents(it)
+            }
+        }
+    }
+
+    abstract fun handleEvents(event: Event)
+
+    protected fun setEffect(builder: () -> Effect) {
+        val effectValue = builder()
+        scope.launch { _effect.send(effectValue) }
+    }
 }
